--- internal/cmd/miniooni/main.go
@@ -1,11 +1,0 @@
-// Command miniooni is a simple binary for research and QA purposes
-// with a CLI interface similar to MK and OONI Probe v2.x.
-package main
-
-//
-// Main function
-//
-
-func main() {
-	Main()
-}
